[minor] Added support for `versions.js` files so you can have a commented config. Fixes #6 [minor] Added support for custom connect plugins. (needs tests) [minor] Started on making some middleware layers optional.
The biggest reason for making these middleware layers optional is that they only add
pointless performance overhead and sometimes even security issues as is in the case
of the REST router. If you don't supply it with a configuration everybody can just
flush your server on demand and see your configuration using the /sync route.